Add LED.blink method - ref #7
This is a rather messy PR, sorry! I got tired of all the repeated
definitions of the read-only pin property so this breaks out that into a
`GPIODevice` base class which `InputDevice` and `OutputDevice` then
inherit from. If you'd rather stick with the repeated definitions let me
know and I'll redo this.