Added PhaseEnableMotor
Moved what I'd originally coded as the Pololu-specific motor into the generic output devices, as a PhaseEnableMotor. This is different to the generic motor in that the two pins aren't both PWM - one is, for the motor power, the other is simply on/off which controls the direction of the output motor.