Make close() nethods more exception-safe
If an exception is thrown during init, it's possible for a close() method (or
one of the methods called by close) to try accessing an attribute that hasn't
been created yet, which would throw an AttributeError exception. This commit
fixes that by ensuring that the close() methods (and methods called by them)
always use hasattr() or getattr() first, before trying to access the
attributes directly.

Fixes #503 and supersedes #494