Re #385, fix #435
This fixes the issue with the DistanceSensor class constantly returning
the maximum distance value. The issue was caused by cab6cc80 which
changed the Pin.when_changed attribute to use weak references to all
handlers. DistanceSensor is the only thing that uses a local function
(not a bound method) as its pin handler. As the pin no longer kept a
reference to it, it disappeared after the constructor was complete and
the echo event never got set.

In addition, it introduces a class-level lock to prevent multiple
distance sensors from interfering with each other (it's the only current
"active" sensor, but this pattern should probably be followed with
future active sensors), and a brief delay between sensor reads to
prevent overlapping echo responses.

Note: this does *not* fix the issue with DistanceSensor operating
remotely over pigpio (which requires a change in the pin handlers); that
is currently scheduled for 1.5 and is why #385 is remaining open for
now.