Fix the Distance Sensor ... again
There's an esoteric edge case in the DistanceSensor; if we miss either a
rising or falling edge the state can persist to the next read. A further
missing edge can result in a huge "over-read". This doesn't affect usage
under pigpio, but under RPi.GPIO (which can miss edges during callbacks)
it's an issue.

Furthermore, the edge case where we return 0 (because we realize we've
missed an edge) harms the overall median when the reflector is very
close to the sensor for long periods. Instead, this patch adjusts the
queue to allow it to "ignore" specified values. When the sensor realizes
it's missed an edge it uses an ignored value which means the sensor's
overall value doesn't suddenly skip to 0.