Fix #600, fix #675
Permit gpiozero to be imported without trying to load a pin factory;
only load a pin factory on first Device construction. While we're at it
re-work the default pin factory loader to avoid pkg_resources import in
as many cases as possible (while keeping it around for third-party pin
implementations). This is just because pkg_resources takes *ages* to
load (especially on an I/O limited platform like the Pi).