I said import, damn it! (#722)
* Fix #600, fix #675

Permit gpiozero to be imported without trying to load a pin factory;
only load a pin factory on first Device construction. While we're at it
re-work the default pin factory loader to avoid pkg_resources import in
as many cases as possible (while keeping it around for third-party pin
implementations). This is just because pkg_resources takes *ages* to
load (especially on an I/O limited platform like the Pi).

* Add get-pip.py to gitignore

Tox creates it for the 3.2 and 3.3 envs

* Support ancient pytest versions

The yield form of pytest fixtures requires @yield_fixture instead of
@fixture in pytest versions <3.0 (which are in turn required to test
python versions 3.2 and 3.3)