More SPI work
Lots of work tidying up SPI. Firstly we ensure that the pin-factory
constructing the SPI interface is always the same one being used to
construct any pins used by the interface (this broke several incorrectly
written tests which are now fixed). Also added the ability for hardware
SPI interfaces to avoid reserving MOSI/MISO pins (as the software
interfaces could).

An example of where this is useful is the Pimoroni breakout-garden SPI
display which uses the MISO pin to select between command/data output
rather than input (as the display doesn't produce any output for the
host so SPI is being used effectively write-only).

This commit also changes the instantiation of hardware SPI interfaces to
use the same pins specification as the software interfaces so now all
SPI interfaces have the same constructor prototype.