Well that was stupid...
Put the bounce_time parameter in the constructor, documented it ...
and then totally forgot to *use* it. Also forgot to set the edges!
This didn't matter with either RPi.GPIO or lgpio which both implicitly
set the edges when when_changed was set, but *did* matter for the
native pin implementation where it didn't work (hurrah for alternative
pin implementations! :)