Renamed SevenSegmentDisplay to LEDCharDisplay
Added a "font" attribute which maps *stuff* to the tuple of LED values
where *stuff* can be a string or basically anything that can key a dict.
This gets slightly complicated when a decimal-point is added as that
pretty much dictates that the value has to be a string (unless we want
to treat the DP *entirely* separately but that makes the value even
messier IMO).

Converted Martin's default 7-seg font to a text-based format and added a
parser for it to tools, then revised that parser to work generically for
any number of segments, and added a default 14-seg font for fun (though
I haven't actually got any hardware to test this :)