Fix \* in explicit constructor declarations
It appears one of the changes from Sphinx 1.x to 3.x was to remove the
necessity for escaping asterisks in constructor overrides.

Unfortunately this makes editing the source in anything with reST syntax
highlighting rather difficult but at least this should be a temporary
problem; the only reason we have the constructor overrides is because we
can't declare the constructors "properly" as the * keyword-only args
syntax doesn't exist in Python 2.x. When we drop support in the next
major release we can remove all the constructor overrides in the docs.