Give type annotations a whirl
Not bad! Fairly impressed with mypy: it picked up a couple of (fairly
obscure) errors with its type analysis (that proc.stdout will be None in
the case that the subprocess fails to launch at all).

That said, on the version of mypy that ships with Ubuntu 20.10 it raises
a false-positive that ConfigParser.default_section doesn't exist (it
does, and has since Python 3.2!).