Move SPI tests
Many of the SPI tests need to be moved under the "real" pins tests as
now that MockFactory descends from PiFactory rather than LocalPiFactory,
most of the logic handling software vs hardware SPI implementations
won't run otherwise.

This is a good opportunity to test hardware SPI initialization properly,
but that does imply that the "regular" pin tests mustn't be sitting on
the SPI pins. While we're at it, might as well ban the I2C pins from
this too in case we get around to adding that interface.

Those tests that don't need to move to real pins, go to test_spi_devices
instead.

Fixes #915