Upgrade pin info to include all pin alt functions
We should now be able to identify all the capabilities of a pin (spi,
i2c, etc.) and next we can upgrade pinout to display alt functions.

Also, rename "spec" to "name". I was never terribly happy with "spec"
for this. "name" is much simpler. Also stripped out the "functions"
dictionary. It really doesn't make sense because in the libgpiod based
future you won't be able to see the pin alt "functions" anyway.

However, we may wish to map specific interfaces to names at some point?