RTD builds "old" projects with sphinx 1.8
See [1] for details of the "feature-flag" that does this. Anyway, turns
out our docs won't build with a sphinx that ancient any more, Sphinx 2.x
is broken on modern Python, as is 3.x, so 4.x is the minimum we now
require.

At least, that's what my local experiments suggest. Let's see if this
works on RTD, or if I need another commit ...

[1]: https://github.com/readthedocs/readthedocs.org/issues/7858#issuecomment-764886373