Support setting custom active_state for InputDevice
Properly supports now normally-closed switches with an internal
pull-up/pull-down resistor as well together with normally-open switches.

Before an exception was raised when attempting to modify the active state
while specifying the `pull_up` parameter. This was rather an artificial
limitation which is now lifted.

Like before, by default it assumed that a button or switch is normally-open
which is a common assumption and keeps the code backwards-compatible.
In this case specifying the `active_state` parameter is not necessary.

Closes #921.